Introduce BGP message recording command
Introduce a `record` subcommand which collects the BGP Open and Update
message from the peer and dump it to the standard output with
human-readable format.

How it works is following:

1. Establishes the TCP connection to the given peer
2. Wait for peer's OPEN message
3. Derive our BGP capability from the peer's OPEN message to let it
   advertise the route they wish to advertise. For example,
   MultiProtocol, ExtendedNextHop, AddPath capabilities are simply
   inherited from the peer (we don't implement any relevant logic for
   them, but we can at least receive it, so safe to advertise that
   capability from our side).
4. Send our OPEN message and establish peering
5. Keep the session alive by sending KEEPALIVE message
6. Capture UPDATE messages and dump it

Note that KEEPALIVE message from the peer will be just ignored and we
don't care about hold timer expiration. Users are responsible for
finish recording by hitting Ctrl-C or sending SIGTERM.

The lifetime of the recording is limited to the single BGP session.
Thus, NOTIFICATION message from the peer or TCP disconnection terminates
the recording immediately.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>