Introduce file recording and report command
Added a `--write` option to the `record` command to allow writing BGP
messages to a file instead of stdout. Currently, the format is just a
raw BGP message binaries. This will be changed in the future.

Also introduced a new `report` command to read and display recorded BGP
messages from a file.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>