recorder,replayer: More fine-grained API
Modify recorder and replayer API to be more flexible. The new API is
inspired by the standard net package's Dialer and Conn API. The recorder
and replayer implements a Dialer, but with different way of BGP session
establishment method (recorder derives OPEN from the peer's one and
replayer replays the one provided by user), and returns the Conn object
which users can read/write BGP messages from/to the already established
BGP session.

Recorder CLI now implements the logic to read from Conn and write it to
the bgpcap file. Replayer CLI takes care of reading the bgpcap file and
write messages to the Conn.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>